+++ src/schemathesis/_hypothesis.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 import asyncio
 import warnings
-from typing import Any, Callable
+from typing import Any, Callable, Optional
 
 import hypothesis
 from hypothesis import Phase
@@ -113,7 +113,7 @@ def add_examples(test: Callable, operation: APIOperation, hook_dispatcher: HookD
     """Add examples to the Hypothesis test, if they are specified in the schema."""
     try:
         examples: list[Case] = [get_single_example(strategy) for strategy in operation.get_strategies_from_examples()]
-    except (OperationSchemaError, HypothesisRefResolutionError, Unsatisfiable):
+    except (OperationSchemaError, HypothesisRefResolutionError, Unsatisfiable) as exc:
         # Invalid schema:
         # In this case, the user didn't pass `--validate-schema=false` and see an error in the output anyway,
         # and no tests will be executed. For this reason, examples can be skipped
@@ -123,6 +123,8 @@ def add_examples(test: Callable, operation: APIOperation, hook_dispatcher: HookD
         # Skipping this exception here allows us to continue the testing process for other operations.
         # Still, we allow running user-defined hooks
         examples = []
+        if isinstance(exc, Unsatisfiable):
+            add_unsatisfied_example_mark(test, exc)
     context = HookContext(operation)  # context should be passed here instead
     GLOBAL_HOOK_DISPATCHER.dispatch("before_add_examples", context, examples)
     operation.schema.hooks.dispatch("before_add_examples", context, examples)
@@ -131,6 +133,18 @@ def add_examples(test: Callable, operation: APIOperation, hook_dispatcher: HookD
     for example in examples:
         test = hypothesis.example(case=example)(test)
     return test
+
+
+def add_unsatisfied_example_mark(test: Callable, exc: Unsatisfiable) -> None:
+    test._schemathesis_unsatisfied_example = exc  # type: ignore
+
+
+def has_unsatisfied_example_mark(test: Callable) -> bool:
+    return hasattr(test, "_schemathesis_unsatisfied_example")
+
+
+def get_unsatisfied_example_mark(test: Callable) -> Optional[Unsatisfiable]:
+    return getattr(test, "_schemathesis_unsatisfied_example", None)
 
 
 def get_single_example(strategy: st.SearchStrategy[Case]) -> Case: